+++ src/vue/modules/loyalty-points/create-invoice-form/store/index.js
@@ -1,4 +1,5 @@
 import { Balance } from '../wrappers/balance'
+import { Movement } from '../wrappers/movement'
 
 import { types } from './types'
 import { api } from '../_api'
@@ -9,6 +10,7 @@ export const state = {
   accountId: '',
   balances: [],
   assetPairs: [],
+  movements: [],
 }
 
 export const mutations = {
@@ -20,6 +22,9 @@ export const mutations = {
   },
   [types.SET_ASSET_PAIRS] (state, assetPairs) {
     state.assetPairs = assetPairs
+  },
+  [types.SET_MOVEMENTS] (state, movements) {
+    state.movements = movements
   },
 }
 
@@ -48,12 +53,38 @@ export const actions = {
 
     commit(types.SET_ASSET_PAIRS, pairs)
   },
+
+  async [types.LOAD_MOVEMENTS] ({ getters, commit }, assetCode) {
+    const balance = getters[types.getBalanceByAssetCode](assetCode)
+
+    if (!balance) {
+      throw new Error(`No balance found for ${assetCode}`)
+    }
+
+    const endpoint = `/${HORIZON_VERSION_PREFIX}/history`
+    const { data: movements } = await api().getWithSignature(endpoint, {
+      page: {
+        order: 'desc',
+      },
+      filter: {
+        account: getters[types.accountId],
+        balance: balance.id,
+      },
+      include: ['effect', 'operation.details'],
+    })
+
+    commit(types.SET_MOVEMENTS, movements)
+  },
 }
 
 export const getters = {
   [types.accountId]: state => state.accountId,
   [types.balances]: state => state.balances.map(b => new Balance(b)),
+  [types.movements]: state => state.movements.map(m => new Movement(m)),
   [types.assetPairs]: state => state.assetPairs,
+  [types.getBalanceByAssetCode]: (_, getters) => assetCode => getters
+    .balances
+    .find(b => b.assetCode === assetCode),
 }
 
 export const createInvoiceFormModule = {